refactor: 파일 구조 변경, 모바일 및 PC 버전 분리
