--- src/component/common/Modal/OperateTime/OperateTime.module.scss
@@ -1,92 +1,0 @@
-.modal {
-  position: fixed;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 100%;
-  left: 0;
-  top: 0;
-  background-color: #00000066;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 430px;
-  height: 536px;
-  background-color: #ffffff;
-
-  &__header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: calc(430px - 32px);
-    height: calc(78px - 48px);
-    background-color: #175c8e;
-    padding: 24px 16px;
-  }
-
-  &__title {
-    font-size: 20px;
-    color: #ffffff;
-  }
-
-  &__close-button {
-    width: 24px;
-    height: 24px;
-    border: 0;
-    background-color: transparent;
-    color: #ffffff;
-    font-size: 100%;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: calc(100% - 64px);
-  height: calc(100% - 64px);
-  padding: 24px 32px 40px;
-}
-
-.table {
-  width: 100%;
-  height: 100%;
-  margin-bottom: 36px;
-  text-align: center;
-
-  th {
-    font-size: 18px;
-    font-weight: 400;
-  }
-
-  &__data {
-    text-align: center;
-  }
-
-  &__time-picker {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-  }
-
-  &__checkbox {
-    appearance: none;
-    border: 1px solid #175c8e;
-    width: 18px;
-    height: 18px;
-    cursor: pointer;
-
-    &:checked {
-      background-color: #175c8e;
-    }
-  }
-}
