+++ src/page/StoreRegistration/component/Modal/OperateTimeMobile/OperateTimeMobile.module.scss
@@ -1,4 +1,4 @@
-.mobile {
+.container {
   position: absolute;
   top: 0;
   left: 0;
