+++ src/page/StoreRegistration/component/Modal/SearchStore/SearchStore.module.scss
@@ -22,8 +22,8 @@
     display: flex;
     align-items: center;
     justify-content: space-between;
-    width: calc(430px - 32px);
-    height: calc(78px - 48px);
+    width: 398px;
+    height: 30px;
     background-color: #175c8e;
     padding: 24px 16px;
   }
@@ -55,8 +55,8 @@
   &__search {
     display: flex;
     align-items: center;
-    width: calc(396px - 27px);
-    height: calc(38px - 18px);
+    width: 369px;
+    height: 20px;
     padding: 9px 11px 9px 16px;
     background-color: #f6f8f9;
     border: 1px solid #d2dae2;
@@ -64,7 +64,7 @@
   }
 
   &__input {
-    width: calc(396px - 27px);
+    width: 369px;
     height: 100%;
     border: none;
     background-color: #f6f8f9;
@@ -86,7 +86,7 @@
   align-items: center;
   justify-content: space-between;
   width: calc(100% - 32px);
-  height: calc(59px - 36px);
+  height: 23px;
   padding: 18px 16px;
   border: 1px solid #d2dae2;
   margin-bottom: 16px;
