+++ src/page/StoreRegistration/view/PC/StoreRegistrationPC.module.scss
